Consolidate all logic for reading/writing filter state into a single file
