+++ ui/src/Tags/MyTagsForm.tsx
@@ -21,7 +21,6 @@ import {connect} from 'react-redux';
 import {GlobalStateProps} from '../Redux/Reducers';
 import {setAllGroupedTagFilterOptionsAction} from '../Redux/Actions';
 import {Tag} from './Tag.interface';
-import {AllGroupedTagFilterOptions} from '../SortingAndFiltering/ProductFilter';
 import {JSX} from '@babel/types';
 import {FilterOption} from '../CommonTypes/Option';
 import {LocationTag} from '../Locations/LocationTag.interface';
@@ -30,6 +29,7 @@ import LocationTags from './LocationTags';
 import ProductTags from './ProductTags';
 
 import '../ModalFormComponents/TagRowsContainer.scss';
+import {AllGroupedTagFilterOptions} from '../SortingAndFiltering/FilterConstants';
 
 // @Todo consolidate (also in MyRolesForm)
 export const INACTIVE_EDIT_STATE_INDEX = -1;
